+++ client/src/commonMain/kotlin/entities/Guild.kt
@@ -5,6 +5,7 @@ import com.serebit.strife.data.*
 import com.serebit.strife.internal.encodeBase64
 import com.serebit.strife.internal.entitydata.GuildData
 import com.serebit.strife.internal.entitydata.GuildMemberData
+import com.serebit.strife.internal.entitydata.GuildMessageChannelData
 import com.serebit.strife.internal.entitydata.toData
 import com.serebit.strife.internal.network.Route
 import com.serebit.strife.internal.packets.BanPacket
@@ -12,7 +13,6 @@ import com.serebit.strife.internal.packets.toIntegration
 import com.serebit.strife.internal.packets.toInvite
 import com.soywiz.klock.DateTimeTz
 import io.ktor.http.isSuccess
-
 
 /**
  * Represents a Guild (aka "server"), or a self-contained community of users. Guilds contain their own
@@ -143,7 +143,7 @@ class Guild internal constructor(private val data: GuildData) : Entity {
 
     /**
      * Create a new [GuildRole]. Set its [name], [permissions], [color], whether it is [mentionable] and whether to
-     * [display it separately in the sidebar][hoist]. Returns the created [GuildRole] if successful, otherwise `null`.
+     * [display it separately in the sidebar][hoist]. Returns the ID of the created role if successful, otherwise `null`.
      * *Requires [Permission.ManageRoles]*
      */
     suspend fun createRole(
@@ -152,9 +152,9 @@ class Guild internal constructor(private val data: GuildData) : Entity {
         color: Color,
         hoist: Boolean = false,
         mentionable: Boolean = false
-    ): Boolean = context.requester.sendRequest(
+    ): Long? = context.requester.sendRequest(
         Route.CreateGuildRole(id, name, permissions.toBitSet(), color.rgb, hoist, mentionable)
-    ).status.isSuccess()
+    ).value?.id
 
     /**
      * Set the Role with id [roleID]'s [position][GuildRole.position].
@@ -192,7 +192,7 @@ class Guild internal constructor(private val data: GuildData) : Entity {
 
     /**
      * Delete [GuildRole] with the given [roleID]. Use this method if only the role ID is available, otherwise the
-     * reccomended method to use is [GuildRole.delete] (though they are functioanlly the same).
+     * recommended method to use is [GuildRole.delete] (though they are functionally the same).
      */
     suspend fun deleteRole(roleID: Long): Boolean =
         context.requester.sendRequest(Route.DeleteGuildRole(id, roleID)).status.isSuccess()
@@ -279,7 +279,7 @@ class Guild internal constructor(private val data: GuildData) : Entity {
      * If [withPruneCount] is set to `true`, this returns the number of [GuildMember]s that would be removed by a
      * [prune] of [days] number of days, or `null` if the request failed or [withPruneCount] is set to `false`.
      *
-     * Note: Discord reccomends setting [withPruneCount] to false for large [Guild]s.
+     * Note: Discord recommends setting [withPruneCount] to false for large [Guild]s.
      *
      * *Defaults [days]=7 and [withPruneCount]=false. Requires [Permission.KickMembers].*
      */
@@ -306,12 +306,21 @@ class Guild internal constructor(private val data: GuildData) : Entity {
     suspend fun deleteIntegration(integrationID: Long): Boolean =
         context.requester.sendRequest(Route.DeleteGuildIntegration(id, integrationID)).status.isSuccess()
 
+    /** Returns the [Guild]'s [AuditLog] or `null` if the request failed. */
+    suspend fun getAuditLog(): AuditLog? = context.requester.sendRequest(Route.GetGuildAuditLog(id, limit = 100))
+        .value?.toAuditLog(data)
+
     /** Returns the [GuildEmbed] for this [Guild] or `null` if the request failed. */
     suspend fun getGuildEmbed(): GuildEmbed? = context.requester.sendRequest(Route.GetGuildEmbed(id)).value
         ?.run { GuildEmbed(this@Guild, enabled, channel_id?.let { getChannel(it) }) }
 
     /** Returns the vanity URL or `null` if not set or the request failed. *Requires [Permission.ManageGuild].* */
     suspend fun getVanityUrl(): String? = context.requester.sendRequest(Route.GetGuildVanityUrl(id)).value?.code
+
+    /** Get all [webhooks][Webhook] of this [Guild]. Returns a [List] of [Webhook], or `null` on failure. */
+    suspend fun getWebhooks(): List<Webhook>? = context.requester.sendRequest(Route.GetChannelWebhooks(id))
+        .value
+        ?.map { it.toEntity(context, data, data.getChannelData(it.channel_id) as GuildMessageChannelData<*, *>) }
 
     companion object {
         /** The minimum character length for a [Guild.name] */