+++ client/src/jvmMain/kotlin/isSubtypeOf.kt
@@ -1,6 +1,0 @@
-package com.serebit.strife.internal
-
-import kotlin.reflect.KType
-import kotlin.reflect.full.isSubtypeOf
-
-internal actual fun KType.isSubtypeOf(other: KType): Boolean = isSubtypeOf(other)